Удалил интерфейс AuthorService, исправил запрос в UserRepository
